configure DCline voltage sense

+++ MotionControl_Firmware_CPU1/Source/src/CPU1_CLA1_common.c
@@ -50,7 +50,7 @@
     #pragma DATA_SECTION(sensorSampleB,"CLADataLS1")
     float32_t sensorSampleB;
     #pragma DATA_SECTION(dcVoltageSense,"CLADataLS1")
-    uint16_t dcVoltageSense;
+    float32_t dcVoltageSense;
 
     // ADC data buffer
     #pragma DATA_SECTION(CLA_SampleBufferA,"CLADataLS1")
